backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
openai==1.3.0
